-refactored places API classes: moved them to one package and created a package for each of them according to their usage -uncommented code for searching user position -adapted code for new API implementation of places (restored places loading on map)
