+++ app/src/main/java/com/travl/guide/mvp/model/network/NetService.java
@@ -3,9 +3,8 @@ package com.travl.guide.mvp.model.network;
 import com.travl.guide.mvp.model.api.articles.ArticleLinksContainer;
 import com.travl.guide.mvp.model.api.city.content.CitiesList;
 import com.travl.guide.mvp.model.api.city.content.CityContent;
-import com.travl.guide.mvp.model.api.newPlaces.Root;
-import com.travl.guide.mvp.model.api.places.ManyPlacesContainer;
-import com.travl.guide.mvp.model.api.places.PlaceContainer;
+import com.travl.guide.mvp.model.api.places.articles.Root;
+import com.travl.guide.mvp.model.api.places.map.ManyPlacesContainer;
 
 import io.reactivex.Single;
 import retrofit2.http.GET;
@@ -15,9 +14,6 @@ import retrofit2.http.Query;
 public interface NetService {
     @GET("api/map/")
     Single<ManyPlacesContainer> loadPlacesForMap(@Query(value = "position", encoded = true) CoordinatesRequest position, @Query("radius") double radius, @Query("detailed") int detailed);
-
-    @GET("api/places/{id}/")
-    Single<PlaceContainer> loadPlace(@Path("id") int id);
 
     @GET("api/query/")
     Single<CityContent> loadCityContent(@Query(value = "position", encoded = true) CoordinatesRequest position);
